Improves article meta, making it more consistent
The old article meta wasn't very consistent, it has icons and text to
describe each element. With this change, it uses only icons and element
titles as hints/tool tips for navigation. Also enable comment count and
moved tags for the end of the article, keeping the header more clean.