Adds downloads statistics
This commit adds a dynamic (through JavaScript) download statistics in
the episode page (right below the episode title). I'm not adding these
statistics to the index because I fear it can degrade the page loading
speed, since I'll be obligated to load the JSON object for every episode
present in the index (10 episodes).